added varification aspx

+++ ComicLibrary/Comic.cs
@@ -12,8 +12,8 @@ namespace ComicLibrary
         private String title;
         private String creator;
         private String description;
-        private int retailPrice;
-        private int resalePrice;
+        private float retailPrice;
+        private float resalePrice;
         private String releaseDate;
         private int sellerid;
         private int buyerid;
@@ -22,8 +22,8 @@ namespace ComicLibrary
         public String Title { get { return title; } set { title = value; } }
         public String Creator { get { return creator; } set { creator = value; } }
         public String Description { get { return description; } set { description = value; } }
-        public int RetailPrice { get { return retailPrice; } set { retailPrice = value; } }
-        public int ResalePrice { get { return resalePrice; } set { resalePrice = value; } }
+        public float RetailPrice { get { return retailPrice; } set { retailPrice = value; } }
+        public float ResalePrice { get { return resalePrice; } set { resalePrice = value; } }
         public String ReleaseDate { get { return releaseDate; } set { releaseDate = value; } }
         public int SellerID { get { return sellerid; } set { sellerid = value; } }
         public int BuyerID { get { return buyerid; } set { buyerid = value; } }
